Refactoring zkp library to specifically longfellow
