+++ multipaz-longfellow/src/commonMain/kotlin/org/multipaz/mdoc/zkp/longfellow/LongfellowZkSystemSpec.kt
@@ -1,0 +1,16 @@
+package org.multipaz.mdoc.zkp.longfellow
+
+/**
+ * Kotlin representation of the native ZkSpecStruct
+ *
+ * @property system the ZK system name.
+ * @property circuitHash the hash of the circuit.
+ * @property numAttributes the number of attributes that the circuit supports.
+ * @property version the version of the ZK spec.
+ */
+data class LongfellowZkSystemSpec(
+    val system: String,
+    val circuitHash: String,
+    val numAttributes: Long,
+    val version: Long
+)
