+++ multipaz-longfellow/src/commonMain/kotlin/org/multipaz/mdoc/zkp/longfellow/ProofGenerationException.kt
@@ -1,0 +1,4 @@
+package org.multipaz.mdoc.zkp.longfellow
+
+/** Exception thrown when the prover fails to generate a proof. */
+class ProofGenerationException(message: String) : Exception(message)
