--- multipaz-zkp/src/commonMain/kotlin/org/multipaz/zkp/ProofGenerationException.kt
@@ -1,4 +1,0 @@
-package org.multipaz.zkp
-
-/** Exception thrown when the prover fails to generate a proof. */
-class ProofGenerationException(message: String) : Exception(message)
