--- multipaz-zkp/src/commonMain/kotlin/org/multipaz/zkp/mdoc/Attribute.kt
@@ -1,14 +1,0 @@
-package org.multipaz.zkp.mdoc
-
-import kotlinx.io.bytestring.ByteString
-
-/**
- * Statements for the zk proof.
- *
- * @property key the key of the attribute.
- * @property value the value of the attribute.
- */
-data class Attribute(
-    val key: String,
-    val value: ByteString
-)
