--- multipaz-zkp/src/commonMain/kotlin/org/multipaz/zkp/mdoc/NativeAttribute.kt
@@ -1,7 +1,0 @@
-package org.multipaz.zkp.mdoc
-
-/** Statements for the zk proof. */
-internal data class NativeAttribute(
-    val key: String,
-    val value: ByteArray
-)
