--- multipaz-zkp/src/commonMain/kotlin/org/multipaz/zkp/mdoc/Proof.kt
@@ -1,14 +1,0 @@
-package org.multipaz.zkp.mdoc
-
-import kotlinx.io.bytestring.ByteString
-
-/**
- * ZK Proof and timestamp used to generate the proof.
- *
- * @property timestamp the ISO formatted timestamp used to generate the proof.
- * @property proof the bytes of the proof.
- * */
-data class Proof (
-    val timestamp: String,
-    val proof: ByteString
-)
