--- multipaz-zkp/src/commonMain/kotlin/org/multipaz/zkp/util/NativeLoader.kt
@@ -1,9 +1,0 @@
-package org.multipaz.zkp.util
-
-/**
- * Loads native libraries
- * Used to allow each system (Android, JVM, iOS) to implement this method per their needs.
- */
-internal expect object NativeLoader {
-    fun loadLibrary(libraryName: String)
-}
