+++ settings.gradle.kts
@@ -62,7 +62,7 @@ include(":multipaz-provisioning-api")
 include(":multipaz-models")
 include(":multipaz-csa")
 include(":multipaz-android-legacy")
-include(":multipaz-zkp")
+include(":multipaz-longfellow")
 include(":multipazctl")
 include(":multipaz-mrtd-reader")
 include(":multipaz-mrtd-reader-android")
